# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AcceptUserActionRequest:
    properties:
      lat:
        example: 41.311081
        type: number
      long:
        example: 69.240562
        type: number
      user_id:
        example: 1
        type: integer
    required:
    - lat
    - long
    - user_id
    type: object
...
  dto.AttendanceAction:
    properties:
      come_status:
        example: 1
        type: integer
      finished_at:
        example: 1737280800
        type: integer
      id:
        example: 1
        type: integer
      leave_branch_id:
        example: 2
        type: integer
      out_status:
        example: 1
        type: integer
      started_at:
        example: 1737277200
        type: integer
      visit_branch_id:
        example: 1
        type: integer
      worked_time:
        example: "08:30"
        type: string
    type: object
...
  dto.CreateUserActionRequest:
    properties:
      come_status:
        example: 1
        type: integer
      finished_at:
        example: 1737280800
        type: integer
      leave_branch_id:
        example: 2
        type: integer
      out_status:
        example: 1
        type: integer
      started_at:
        example: 1737277200
        type: integer
      user_id:
        example: 1
        type: integer
      visit_branch_id:
        example: 1
        type: integer
    required:
    - come_status
    - started_at
    - user_id
    - visit_branch_id
    type: object
  dto.CreateUserRequest:
    properties:
      department_id:
        example: 1
        type: integer
      full_name:
        example: John Doe
        type: string
      password:
        example: securepassword
        type: string
      phone:
        example: "+998901234567"
        type: string
      role:
        example: 1
        type: integer
      schedule_id:
        example: 1
        type: integer
      section_id:
        example: 1
        type: integer
      tg_chat_id:
        example: 123456789
        type: integer
    required:
    - department_id
    - full_name
    - password
    - phone
    - role
    - section_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: department not found
        type: string
    type: object
...
  dto.UpdateUserActionRequest:
    properties:
      come_status:
        example: 1
        type: integer
      finished_at:
        example: 1737280800
        type: integer
      leave_branch_id:
        example: 2
        type: integer
      out_status:
        example: 1
        type: integer
      started_at:
        example: 1737277200
        type: integer
      user_id:
        example: 1
        type: integer
      visit_branch_id:
        example: 1
        type: integer
    required:
    - come_status
    - started_at
    - user_id
    - visit_branch_id
    type: object
  dto.UpdateUserRequest:
    properties:
      department_id:
        example: 1
        type: integer
      full_name:
        example: John Doe
        type: string
      joined_at:
        example: 1737277200
        type: integer
      phone:
        example: "+998901234567"
        type: string
      role:
        example: 1
        type: integer
      schedule_id:
        example: 1
        type: integer
      section_id:
        example: 1
        type: integer
      tg_chat_id:
        example: 123456789
        type: integer
    required:
    - department_id
    - full_name
    - joined_at
    - phone
    - role
    - section_id
    type: object
  dto.UserActionResponse:
    properties:
      come_status:
        example: 1
        type: integer
      finished_at:
        example: 1737280800
        type: integer
      id:
        example: 1
        type: integer
      leave_branch_id:
        example: 2
        type: integer
      out_status:
        example: 1
        type: integer
      started_at:
        example: 1737277200
        type: integer
      user_id:
        example: 1
        type: integer
      visit_branch_id:
        example: 1
        type: integer
    type: object
  dto.UserAttendanceResponse:
    properties:
      actions:
        items:
          $ref: '#/definitions/dto.AttendanceAction'
        type: array
      total_worked_time:
        example: "16:30"
        type: string
      user_full_name:
        example: John Doe
        type: string
      user_id:
        example: 1
        type: integer
      user_phone:
        example: "+998901234567"
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        example: 1737277200
        type: integer
      department_id:
        example: 1
        type: integer
      full_name:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      joined_at:
        example: 1737277200
        type: integer
      phone:
        example: "+998901234567"
        type: string
      role:
        example: 1
        type: integer
      schedule_id:
        example: 1
        type: integer
      section_id:
        example: 1
        type: integer
      tg_chat_id:
        example: 123456789
        type: integer
      updated_at:
        example: 1737277200
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing departments in Sahiy Management system
  title: Sahiy Management API
  version: "1.0"
paths:
...
  /user-actions:
    get:
      description: Retrieve a list of all user actions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserActionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all user actions
      tags:
      - user-actions
    post:
      consumes:
      - application/json
      description: Create a new user action with the provided data
      parameters:
      - description: User action data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserActionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new user action
      tags:
      - user-actions
  /user-actions/{id}:
    delete:
      description: Delete a user action by ID
      parameters:
      - description: User action ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a user action
      tags:
      - user-actions
    get:
      description: Retrieve a single user action by its ID
      parameters:
      - description: User action ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a user action by ID
      tags:
      - user-actions
    put:
      consumes:
      - application/json
      description: Update an existing user action by ID
      parameters:
      - description: User action ID
        in: path
        name: id
        required: true
        type: integer
      - description: User action data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a user action
      tags:
      - user-actions
  /user-actions/accept:
    post:
      consumes:
      - application/json
      description: Check-in or check-out based on user's geo-location relative to
        branches
      parameters:
      - description: User location data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AcceptUserActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserActionResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Accept a user action (check-in/check-out)
      tags:
      - user-actions
  /user-actions/attendance:
    get:
      description: 'Get attendance reports for a branch within a date range (Format:
        d.m.Y)'
      parameters:
      - description: Branch ID (0 for all)
        in: query
        name: branch_id
        type: integer
      - description: 'From Date (Format: d.m.Y, e.g. 24.01.2026)'
        in: query
        name: from_date
        required: true
        type: string
      - description: 'To Date (Format: d.m.Y, e.g. 25.01.2026)'
        in: query
        name: to_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserAttendanceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get attendance reports
      tags:
      - user-actions
...
  /users:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided data
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a user
      tags:
      - users
    get:
      description: Retrieve a single user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"